lock withLock 参数 millisecondsToTimeout 由 long 改为 int

+++ src/main/java/com/jfinal/plugin/redis/Cache.java
@@ -1425,7 +1425,7 @@ public class Cache {
 	 * @param millisecondsToTimeout 获取锁的超时时间，单位毫秒
 	 * @return 获取成功返回字符串，否则返回 false。释放锁时需要用到返回的字符串
 	 */
-	public String lock(String name, int secondsToExpire, long millisecondsToTimeout) {
+	public String lock(String name, int secondsToExpire, int millisecondsToTimeout) {
 		Jedis jedis = getJedis();
 		try {
 			String lockId = java.util.UUID.randomUUID().toString();
@@ -1468,7 +1468,7 @@ public class Cache {
 	 * 		// 业务操作代码
 	 * });
 	 */
-	public boolean withLock(String name, int secondsToExpire, long millisecondsToTimeout, com.jfinal.kit.Func.F00 fun) {
+	public boolean withLock(String name, int secondsToExpire, int millisecondsToTimeout, com.jfinal.kit.Func.F00 fun) {
 		String lockId = lock(name, secondsToExpire, millisecondsToTimeout);
 		if (lockId == null) {
 			return false;
